# docs/openapi/core.yaml
openapi: 3.1.0
info:
  title: Tilecast API
  version: 0.10.0
  license:
    name: AGPL-3.0-only
  description: >-
    Management authentication accepts either an enrolled dashboard session
    cookie or a bearer user grant (an OAuth access token or a personal
    access token), and both produce the same user principal with the live
    role and screen scope. Grants additionally carry a scope ceiling of
    read, write, or admin (admin implies write implies read): reads take
    read, mutations take write, and user, token, grant, backup, and
    security administration take admin. CSRF is cookie-only: unsafe
    browser requests require X-CSRF-Token, bearer requests never send it.
    Browser security ceremonies (MFA, passkeys, logout, OAuth approval)
    stay session-only.
paths:
...
  /api/v1/me/security/grants:
    get:
      operationId: listOAuthGrants
      security: [{ sessionCookie: [] }, { userBearer: [] }]
      description: >-
        Requires an authenticated user with the admin scope. Lists the
        caller's authorization grants newest first, revoked included.
      responses:
        "200": { description: Grant list }
        "401": { description: Dashboard authentication required }
  /api/v1/me/security/grants/{id}:
    delete:
      operationId: revokeOAuthGrant
      description: >-
        Requires an authenticated user with the admin scope. Cookie
        requests additionally require the X-CSRF-Token header. Revokes
        one of the caller's grants and every token under it. Personal
        access tokens are revoked through this same endpoint.
      parameters:
        - { $ref: "#/components/parameters/ResourceID" }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Grant revoked }
        "401": { description: Dashboard authentication required }
        "403": { description: CSRF token required }
        "404": { description: Grant unknown }
...
components:
  parameters:
...
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    CSRFToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
      description: Cookie-backed browser requests only. Bearer grants never send it.
  securitySchemes:
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: Tilecast device credential
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: Tilecast user grant (tca_ OAuth access token or tcp_ personal access token)
    sessionCookie:
      type: apiKey
      in: cookie
      name: tilecast_session
      description: Enrolled dashboard session cookie
...